Implement cobra cli argument parser
